+++ src/Montray/ServiceManagement/SensorServiceState.cs
@@ -1,0 +1,11 @@
+namespace Montray.ServiceManagement;
+
+public enum SensorServiceState
+{
+    NotInstalled,
+    Stopped,
+    StartPending,
+    Running,
+    StopPending,
+    Unknown
+}